support synonyms test

+++ main.py
@@ -22,32 +22,53 @@ def main():
     world = worlds.load(config)
     model = models.load(config)
     trainer = trainers.load(config)
-    trainer.train(model, world)
+    if config.synonyms is not None:
+        trainer.evaluate(model, world)
+    else:
+        trainer.train(model, world)
 
 def configure():
     # load config
     with open("config.yaml") as config_f:
         config = Struct(**yaml.load(config_f, Loader=yaml.FullLoader))
 
-    # set up experiment
-    config.experiment_dir = os.path.join("experiments/%s" % config.name)
-    if os.path.exists(config.experiment_dir):
-        shutil.rmtree(config.experiment_dir)
-    assert not os.path.exists(config.experiment_dir), \
-            "Experiment %s already exists!" % config.experiment_dir
-    os.mkdir(config.experiment_dir)
+    if config.synonyms is not None:
+        # set up experiment
+        config.experiment_dir = os.path.join("experiments/%s" % config.name)
+        assert os.path.exists(config.experiment_dir), \
+                "Experiment %s does not exist!" % config.experiment_dir
 
-    # set up logging
-    log_name = os.path.join(config.experiment_dir, "run.log")
-    logging.basicConfig(filename=log_name, level=logging.DEBUG,
-            format='%(asctime)s %(levelname)-8s %(message)s')
-    def handler(type, value, tb):
-        logging.exception("Uncaught exception: %s", str(value))
-        logging.exception("\n".join(traceback.format_exception(type, value, tb)))
-    sys.excepthook = handler
+        # set up logging
+        log_name = os.path.join(config.experiment_dir, "{}.log".format(config.synonyms))
+        logging.basicConfig(filename=log_name, level=logging.DEBUG,
+                format='%(asctime)s %(levelname)-8s %(message)s')
+        def handler(type, value, tb):
+            logging.exception("Uncaught exception: %s", str(value))
+            logging.exception("\n".join(traceback.format_exception(type, value, tb)))
+        sys.excepthook = handler
 
-    logging.info("BEGIN")
-    logging.info(str(config))
+        logging.info("BEGIN")
+        logging.info(str(config))
+    else:
+        # set up experiment
+        config.experiment_dir = os.path.join("experiments/%s" % config.name)
+        if os.path.exists(config.experiment_dir):
+            shutil.rmtree(config.experiment_dir)
+        assert not os.path.exists(config.experiment_dir), \
+                "Experiment %s already exists!" % config.experiment_dir
+        os.mkdir(config.experiment_dir)
+
+        # set up logging
+        log_name = os.path.join(config.experiment_dir, "run.log")
+        logging.basicConfig(filename=log_name, level=logging.DEBUG,
+                format='%(asctime)s %(levelname)-8s %(message)s')
+        def handler(type, value, tb):
+            logging.exception("Uncaught exception: %s", str(value))
+            logging.exception("\n".join(traceback.format_exception(type, value, tb)))
+        sys.excepthook = handler
+
+        logging.info("BEGIN")
+        logging.info(str(config))
 
     return config
